Generalize 'transpose' function and extract it into separate module
